refactor(renderer): isolate integration display model and controller [skip ci]
