+++ desktop/renderer/features/integration-center/index.mjs
@@ -1,0 +1,2 @@
+export { adapterView, previewView } from './model.mjs';
+export { createIntegrationCenter } from './controller.mjs';
