+++ desktop/renderer/index.html
@@ -735,7 +735,7 @@
   <script defer="defer" src="campus-service-workspace.js"></script>
   <script defer="defer" src="proxy-auth-migration.js"></script>
   <script defer="defer" src="school-profile-selector.js"></script>
-  <script defer="defer" src="integration-center.js"></script>
+  <script type="module" src="integration-center.js"></script>
   <script type="module" src="app.js"></script>
   <p id="cardBoardLiveRegion" class="visually-hidden" aria-live="polite"></p>
 </body>
